+++ CHANGELOG.md
@@ -121,6 +121,7 @@ CRA Templates has been created for following examples.
 
 ### Fixed - 2020-10-01
 - examples/apollo-real-time-chart
+- examples/lattice-ssr
 - examples: fix apollo-real-time-chart example
 
 [unreleased]: https://github.com/:latticejs/lattice/compare/v1.0.1-beta.1...HEAD